moving files around, pipe tests
