+++ colossus/src/main/scala/colossus/core/DataBuffer.scala
@@ -12,7 +12,7 @@ import java.nio.channels.SocketChannel
  */
 sealed trait DataReader
 
-case class DataStream(sink: service.Sink[DataBuffer]) extends DataReader
+case class DataStream(sink: controller.Sink[DataBuffer]) extends DataReader
 
 /** A thin wrapper around a NIO ByteBuffer with data to read
  *
@@ -21,7 +21,7 @@ case class DataStream(sink: service.Sink[DataBuffer]) extends DataReader
  * read from once and cannot be reset.
  * 
  */
-case class DataBuffer(data: ByteBuffer) extends DataReader with service.Copyable[DataBuffer]{
+case class DataBuffer(data: ByteBuffer) extends DataReader {
   /** Get the next byte, removing it from the buffer
    *
    * WARNING : This method will throw an exception if no data is left.  It is
@@ -54,8 +54,6 @@ case class DataBuffer(data: ByteBuffer) extends DataReader with service.Copyable
    * @return a new DataBuffer containing only the unread data in this buffer
    */
   def takeCopy: DataBuffer = DataBuffer.fromByteString(ByteString(takeAll))
-
-  def copy = takeCopy
 
   /** Directly copy data into a target byte array
    * @param buffer the array to copy into
