+++ colossus/src/main/scala/colossus/service/ServiceClient.scala
@@ -1,10 +1,12 @@
 package colossus
 package service
 
+
 import java.net.InetSocketAddress
 
 import akka.actor._
 import akka.event.Logging
+import colossus.controller._
 import colossus.core._
 import colossus.metrics._
